Numerous and boring changes for 32-bit long support on Windows/amd64. Warning! Warning! Mass replacement happened. There could be new bugs.

+++ src/runtime/alloc.c
@@ -158,7 +158,7 @@ alloc_cons(lispobj car, lispobj cdr)
 }
 
 lispobj
-alloc_number(long n)
+alloc_number(intptr_t n)
 {
     struct bignum *ptr;
