+++ src/runtime/alloc.h
@@ -21,7 +21,7 @@ extern lispobj *general_alloc(sword_t bytes, int page_type_flag);
 #endif
 
 extern lispobj alloc_cons(lispobj car, lispobj cdr);
-extern lispobj alloc_number(long n);
+extern lispobj alloc_number(intptr_t n);
 extern lispobj alloc_string(char *str);
 extern lispobj alloc_sap(void *ptr);
 extern lispobj alloc_base_string(char *str);
